Clean up after large merge
