+++ hpx/runtime/threads/policies/shared_priority_queue_scheduler.hpp
@@ -99,10 +99,10 @@ namespace hpx { namespace threads { namespace policies {
 ///////////////////////////////////////////////////////////////////////////
 #if defined(HPX_HAVE_CXX11_STD_ATOMIC_128BIT)
     using default_shared_priority_queue_scheduler_terminated_queue =
-        lockfree_fifo;
+        lockfree_lifo; // concurrentqueue_fifo; // lockfree_lifo;
 #else
     using default_shared_priority_queue_scheduler_terminated_queue =
-        lockfree_fifo;
+        lockfree_fifo; // concurrentqueue_fifo; // lockfree_fifo;
 #endif
 
     // Holds core/queue ratios used by schedulers.
@@ -127,7 +127,7 @@ namespace hpx { namespace threads { namespace policies {
     /// addition, the shared_priority_queue_scheduler is NUMA-aware and takes
     /// NUMA scheduling hints into account when creating and scheduling work.
     template <typename Mutex = std::mutex,
-        typename PendingQueuing = lockfree_fifo,
+        typename PendingQueuing = lockfree_fifo, // concurrentqueue_fifo,
         typename TerminatedQueuing =
             default_shared_priority_queue_scheduler_terminated_queue>
     class shared_priority_queue_scheduler : public scheduler_base
@@ -143,7 +143,7 @@ namespace hpx { namespace threads { namespace policies {
             init_parameter(std::size_t num_worker_threads,
                 core_ratios cores_per_queue,
                 detail::affinity_data const& affinity_data,
-                thread_queue_init_parameters thread_queue_init = {},
+                const thread_queue_init_parameters &thread_queue_init,
                 char const* description = "shared_priority_queue_scheduler")
               : num_worker_threads_(num_worker_threads)
               , cores_per_queue_(cores_per_queue)
@@ -689,7 +689,7 @@ namespace hpx { namespace threads { namespace policies {
                 {
                     q_counts_[d] = core_use_[d].size();
                     // init with {cores, queues} on this domain
-                    numa_holder_[d].init(q_counts_[d], thread_queue_init_);
+                    numa_holder_[d].init(q_counts_[d]);
                 }
 
                 debug::output("p_lookup_  ", &p_lookup_[0],  &p_lookup_[num_workers_]);
@@ -743,7 +743,7 @@ namespace hpx { namespace threads { namespace policies {
                     if (local_q % cores_per_queue_.high_priority == 0)
                     {
                         // if we will own the queue, create it
-                        hp_queue = new thread_queue_type(queue_parameters_, nullptr, local_q);
+                        hp_queue = new thread_queue_type(queue_parameters_, local_q);
                         owner_mask |= 1;
                     }
                     else {
@@ -755,7 +755,7 @@ namespace hpx { namespace threads { namespace policies {
                 // Normal priority
                 if (local_q % cores_per_queue_.normal_priority == 0) {
                     // if we will own the queue, create it
-                    np_queue = new thread_queue_type(queue_parameters_, nullptr, local_q);
+                    np_queue = new thread_queue_type(queue_parameters_, local_q);
                     owner_mask |= 2;
                 }
                 else {
@@ -767,7 +767,7 @@ namespace hpx { namespace threads { namespace policies {
                 if (cores_per_queue_.low_priority>0) {
                     if (local_q % cores_per_queue_.low_priority == 0) {
                         // if we will own the queue, create it
-                        lp_queue = new thread_queue_type(queue_parameters_, nullptr, local_q);
+                        lp_queue = new thread_queue_type(queue_parameters_, local_q);
                         owner_mask |= 4;
                     }
                     else {
